Run the queries and update gauges in the background
It's quite possible for the battery of queries to take longer than
Prometheus is prepared to wait for its scrape request to
complete. Since the queries are run in the HTTP handler, when that
happens, Prometheus either gets a 500 response or times out, and
considers the pod as down.

Instead of running queries on every scrape, run them to a
schedule ("as often as every five seconds").

(NB there is now no point in testing that the handler returns a 500)